requirements.txt
pandas
sentence-transformers
faiss-cpu
numpy
tqdm
openpyxl
